+++ templates/jefe-cli.sh
@@ -1,0 +1,8 @@
+#!/bin/bash
+#
+# custom tasks jefe-cli.sh
+#
+
+# Load utilities
+source ~/.jefe-cli/libs/utilities.sh
+