+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+# Version 1.3.1
+- module: Fix error do not up nginx to restart proyect for laravel and php-nginx-mysql module.
+- module: Fix do not working commands ps, restart, logs
+- core: Remove start command
+- core: Set detached mode as only option in up command
+- core: Add option ruby on rails in selected option
+# Version 1.3.0
+- module: Add select containers in itbash command for laravel module.
+- fix: Fix error do not found host in nginx by laravel module.
+- module: Add Ruby On Rails module.
+- module: Add Laravel module.
+- module: Dynamically establish the name of container services.
+- module: Active mod_rewrite to php-apache-mysql image.
 # Version 1.2.0
 - module: Add PHP-Apache-MySQL module.
 - fix: Set correct container name of mysql to php-mysql-nginx module.